refactor(multimedia): Moved persistense directory into infrastructure
